requirements.txt
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.2
pylint==2.17.5
coverage==7.3.1
httpx==0.24.0
pyspellchecker==0.7.0
Flask==2.2.2
Werkzeug==2.2.2

